refactor: architectural improvements — eliminate god components and duplicated patterns
- Extract MovieSchema.from_model() classmethod to consolidate ORM-to-schema mapping across 3 routers
- Add require_consent FastAPI dependency to replace duplicated consent guards
- Extract should_suggest_swipe() helper to centralize threshold logic and constants
- Split auth.py: move user profile/settings/consent/sync endpoints to new routers/users.py
- Move background task wrappers (sync_pool, backfill_posters) to service layer
- Reduce auth.py from 507 to ~250 lines, improving SRP and reducing cognitive load

These changes reduce code duplication (84 lines removed vs 231 added net), improve testability
of business logic, and make future changes to authentication, user management, and background
tasks less error-prone by isolating concerns.
